+++ NEWS.md
@@ -3,6 +3,7 @@
 * Added a `NEWS.md` file to track changes to the package.
 * Added an energy cost linear optimization
 * Smart charging function has new parameters to select the optimization objective and the scheduling method.
+* Added parallel processing in the optimization functions